Rename feature to match with new PR title
