+++ source/refined-github.ts
@@ -214,7 +214,7 @@ import './features/rgh-sponsor-button';
 import './features/rgh-linkify-features';
 import './features/collapse-markdown-sections';
 import './features/conversation-activity-filter';
-import './features/select-all-notifications';
+import './features/select-all-notifications-shortcut';
 import './features/no-duplicate-list-update-time';
 
 // Add global for easier debugging
